Equation Processor, using Stack & CusorArray
My computer science, data structure, phase 2 project. An equation processor that takes a file containing tags and delimiters as an input. The program validates the file first (tags and delimiters) then converts from infix to postfix and from postfix to prefix.  But with the additional challenge that we can only use one cursor array for the whole program and within that one cursor array create as many stacks as needed, also the cursor array has to double the size every time its capacity is full.